Organize onion modules and processor tests
