+++ crates/node/src/native/mod.rs
@@ -1,4 +1,3 @@
 pub mod cli;
 pub mod config;
 pub mod endpoint;
-pub mod onion_http_proxy;
