+++ crates/node/src/onion/proxy/http/mod.rs
@@ -3,30 +3,95 @@
 
 use std::net::SocketAddr;
 use std::sync::Arc;
+use std::time::Duration;
 
 use tokio::io::AsyncReadExt;
 use tokio::io::AsyncWriteExt;
 use tokio::net::TcpListener;
 use tokio::net::TcpStream;
+use tokio::sync::Semaphore;
+use tokio::time::timeout;
 
+use super::OnionProxyConfig;
+use super::OnionProxyTarget;
 use crate::error::Error;
 use crate::error::Result;
 use crate::onion::tcp::NativeOnionCircuitHandle;
-use crate::onion_proxy::OnionProxyConfig;
-use crate::onion_proxy::OnionProxyTarget;
+use crate::onion::OnionServiceName;
 use crate::processor::Processor;
 
 const MAX_CONNECT_HEADER_BYTES: usize = 8192;
+/// Default deadline for receiving a complete CONNECT header.
+pub const DEFAULT_CONNECT_HEADER_TIMEOUT_SECS: u64 = 10;
+/// Default concurrent connection bound for the native CONNECT ingress.
+pub const DEFAULT_MAX_CONNECT_CONNECTIONS: usize = 1024;
+
+/// Return the default CONNECT header deadline in seconds.
+pub const fn default_connect_header_timeout_secs() -> u64 {
+    DEFAULT_CONNECT_HEADER_TIMEOUT_SECS
+}
+
+/// Return the default native CONNECT ingress concurrency bound.
+pub const fn default_max_connect_connections() -> usize {
+    DEFAULT_MAX_CONNECT_CONNECTIONS
+}
 
 /// Runtime options for the native onion HTTP CONNECT proxy.
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub struct OnionHttpProxyOptions {
     /// Local bind address.
     pub listen_addr: SocketAddr,
+    /// TCP onion-exit service used for local CONNECT requests.
+    pub service: OnionServiceName,
     /// Desired hop count including the exit. `0` uses node default.
     pub hop_count: usize,
     /// Whether route selection may use fewer hops when too few relays are live.
     pub allow_short_paths: bool,
+    /// Maximum concurrent local CONNECT requests accepted by this ingress.
+    pub max_connections: usize,
+    /// Deadline for receiving a complete CONNECT header.
+    pub header_timeout: Duration,
+}
+
+impl OnionHttpProxyOptions {
+    /// Build options with production defaults for resource bounds.
+    pub fn new(
+        listen_addr: SocketAddr,
+        service: OnionServiceName,
+        hop_count: usize,
+        allow_short_paths: bool,
+    ) -> Self {
+        Self {
+            listen_addr,
+            service,
+            hop_count,
+            allow_short_paths,
+            max_connections: DEFAULT_MAX_CONNECT_CONNECTIONS,
+            header_timeout: Duration::from_secs(DEFAULT_CONNECT_HEADER_TIMEOUT_SECS),
+        }
+    }
+
+    fn validate(&self) -> Result<()> {
+        if self.max_connections == 0 {
+            return Err(Error::InvalidConfig(
+                "onion_http_proxy_max_connections must be greater than zero".to_string(),
+            ));
+        }
+        if self.header_timeout.is_zero() {
+            return Err(Error::InvalidConfig(
+                "onion_http_proxy_header_timeout_secs must be greater than zero".to_string(),
+            ));
+        }
+        self.proxy_config().map(|_| ())
+    }
+
+    fn proxy_config(&self) -> Result<OnionProxyConfig> {
+        OnionProxyConfig::tcp_connect_service(
+            self.service.clone(),
+            self.hop_count,
+            self.allow_short_paths,
+        )
+    }
 }
 
 /// Run a native HTTP CONNECT proxy for onion TCP exits.
@@ -35,6 +100,7 @@ pub async fn run_onion_http_proxy(
     processor: Arc<Processor>,
     onion: NativeOnionCircuitHandle,
 ) -> Result<()> {
+    options.validate()?;
     let listener = TcpListener::bind(options.listen_addr)
         .await
         .map_err(|error| Error::OnionProxyIoError(format!("bind HTTP proxy listener: {error}")))?;
@@ -42,8 +108,14 @@ pub async fn run_onion_http_proxy(
         Error::OnionProxyIoError(format!("read HTTP proxy listener address: {error}"))
     })?;
     println!("Onion HTTP CONNECT proxy endpoint: http://{listen_addr}");
+    let permits = Arc::new(Semaphore::new(options.max_connections));
 
     loop {
+        let permit = permits
+            .clone()
+            .acquire_owned()
+            .await
+            .map_err(|_| Error::Lock)?;
         let (stream, peer_addr) = listener.accept().await.map_err(|error| {
             Error::OnionProxyIoError(format!("accept HTTP proxy connection: {error}"))
         })?;
@@ -51,6 +123,7 @@ pub async fn run_onion_http_proxy(
         let onion = onion.clone();
         let options = options.clone();
         tokio::spawn(async move {
+            let _permit = permit;
             if let Err(error) = handle_connect(stream, processor, onion, options).await {
                 tracing::warn!("onion HTTP proxy request from {peer_addr} failed: {error:?}");
             }
@@ -64,7 +137,7 @@ async fn handle_connect(
     onion: NativeOnionCircuitHandle,
     options: OnionHttpProxyOptions,
 ) -> Result<()> {
-    let target = match read_connect_target(&mut stream).await {
+    let target = match read_connect_target(&mut stream, options.header_timeout).await {
         Ok(target) => target,
         Err(error) => {
             let _ = write_proxy_response(&mut stream, "400 Bad Request").await;
@@ -72,10 +145,7 @@ async fn handle_connect(
         }
     };
     let proxy_route = processor
-        .build_onion_proxy_route(
-            OnionProxyConfig::tcp_connect(options.hop_count, options.allow_short_paths),
-            target,
-        )
+        .build_onion_proxy_route(options.proxy_config()?, target)
         .await?;
     let opened = onion
         .open_tcp_stream(proxy_route.route, proxy_route.target)
@@ -85,8 +155,18 @@ async fn handle_connect(
     Ok(())
 }
 
-async fn read_connect_target(stream: &mut TcpStream) -> Result<OnionProxyTarget> {
-    let header = read_http_header(stream).await?;
+async fn read_connect_target(
+    stream: &mut TcpStream,
+    header_timeout: Duration,
+) -> Result<OnionProxyTarget> {
+    let header = timeout(header_timeout, read_http_header(stream))
+        .await
+        .map_err(|_| {
+            Error::HttpRequestError(format!(
+                "HTTP CONNECT header timed out after {} ms",
+                header_timeout.as_millis()
+            ))
+        })??;
     let header = std::str::from_utf8(&header)
         .map_err(|_| Error::HttpRequestError("HTTP CONNECT header is not UTF-8".to_string()))?;
     let request_line = header
@@ -158,8 +238,20 @@ async fn write_proxy_response(stream: &mut TcpStream, status: &str) -> Result<()
 }
 
 #[cfg(test)]
-mod tests {
+mod tests;
+
+#[cfg(test)]
+mod unit_tests {
     use super::*;
+
+    fn options() -> OnionHttpProxyOptions {
+        OnionHttpProxyOptions::new(
+            SocketAddr::from(([127, 0, 0, 1], 0)),
+            OnionServiceName::tcp(),
+            0,
+            false,
+        )
+    }
 
     #[test]
     fn connect_request_line_parses_target() -> Result<()> {
@@ -174,6 +266,34 @@ mod tests {
         assert!(matches!(
             parse_connect_request_line("GET http://example.com/ HTTP/1.1"),
             Err(Error::HttpRequestError(_))
+        ));
+    }
+
+    #[test]
+    fn proxy_options_build_custom_tcp_service_config() -> Result<()> {
+        let mut options = options();
+        options.service = OnionServiceName::parse("web")?;
+
+        let proxy = options.proxy_config()?;
+
+        assert_eq!(proxy.exit_service(), "web");
+        Ok(())
+    }
+
+    #[test]
+    fn proxy_options_reject_unbounded_connection_model() {
+        let mut zero_connections = options();
+        zero_connections.max_connections = 0;
+        assert!(matches!(
+            zero_connections.validate(),
+            Err(Error::InvalidConfig(_))
+        ));
+
+        let mut zero_timeout = options();
+        zero_timeout.header_timeout = Duration::ZERO;
+        assert!(matches!(
+            zero_timeout.validate(),
+            Err(Error::InvalidConfig(_))
         ));
     }
 }
