+++ crates/node/src/prelude.rs
@@ -27,6 +27,12 @@ pub use self::rings_core::session::SessionSk;
 pub use self::rings_core::session::SessionSkBuilder;
 pub use self::rings_core::swarm::Swarm;
 pub use self::rings_core::swarm::SwarmBuilder;
+pub use crate::onion::proxy::OnionProxyConfig;
+pub use crate::onion::proxy::OnionProxyProtocol;
+pub use crate::onion::proxy::OnionProxyRoute;
+pub use crate::onion::proxy::OnionProxyTarget;
+pub use crate::onion::proxy::ONION_PROXY_HTTPS_SERVICE;
+pub use crate::onion::proxy::ONION_PROXY_TCP_SERVICE;
 pub use crate::onion::OnionExitDescriptor;
 pub use crate::onion::OnionExitPolicy;
 pub use crate::onion::OnionExitRegistration;
@@ -36,12 +42,6 @@ pub use crate::onion::OnionRoute;
 pub use crate::onion::OnionRouteRequest;
 pub use crate::onion::ONION_EXITS_TOPIC;
 pub use crate::onion::ONION_RELAY_CAPABILITY;
-pub use crate::onion_proxy::OnionProxyConfig;
-pub use crate::onion_proxy::OnionProxyProtocol;
-pub use crate::onion_proxy::OnionProxyRoute;
-pub use crate::onion_proxy::OnionProxyTarget;
-pub use crate::onion_proxy::ONION_PROXY_HTTPS_SERVICE;
-pub use crate::onion_proxy::ONION_PROXY_TCP_SERVICE;
 pub use crate::online::OnlineNodeDescriptor;
 pub use crate::online::OnlineNodeType;
 pub use crate::online::ONLINE_NODES_TOPIC;
